Rewrite env_pass_validator to be a value parser
This rewrites the environment value parsing code using the previous
implementation of the CLI validator function, that had a more
sophisticated parsing implementation.

The CLI validator function gets then transformed to a value parser,
using that parsing code again.

Less code, less bugs. Yay!
