+++ src/commands/tree_of.rs
@@ -46,11 +46,10 @@ pub async fn tree_of(
         .map(ImageName::from);
 
     let additional_env = matches
-        .get_many::<String>("env")
+        .get_many::<(EnvironmentVariableName, String)>("env")
         .unwrap_or_default()
-        .map(AsRef::as_ref)
-        .map(crate::util::env::parse_to_env)
-        .collect::<Result<Vec<(EnvironmentVariableName, String)>>>()?;
+        .map(Clone::clone)
+        .collect::<Vec<(EnvironmentVariableName, String)>>();
 
     let condition_data = ConditionData {
         image_name: image_name.as_ref(),
